Changed peak curvature to find the maximum curvature numerically
Changed peak curvature to find the maximum curvature numerically by evaluating curvature along the interval (0,1). Also changed peak curvature to select the curvature maximum closest to t=0.5 to prevent areas near endpoints from be selected as having the peak curvature, even though it may be technically true we are interested most in the distal region of the phragmoplast.